+++ src/internal/account.ts
@@ -52,10 +52,9 @@ import {
   GetAccountTransactionsCount,
 } from "../types/generated/queries";
 import { memoizeAsync } from "../utils/memoize";
-import { Secp256k1PrivateKey, AuthenticationKey, Ed25519PrivateKey, createObjectAddress } from "../core";
+import { Secp256k1PrivateKey, AuthenticationKey, Ed25519PrivateKey } from "../core";
 import { CurrentFungibleAssetBalancesBoolExp } from "../types/generated/types";
 import { getTableItem } from "./table";
-import { APTOS_COIN } from "../utils";
 
 export async function getInfo(args: {
   aptosConfig: AptosConfig;
@@ -400,58 +399,6 @@ export async function getAccountTransactionsCount(args: {
   // commonjs (aka cjs) doesnt handle Nullish Coalescing for some reason
   // might be because of how ts infer the graphql generated scheme type
   return data.account_transactions_aggregate.aggregate ? data.account_transactions_aggregate.aggregate.count : 0;
-}
-
-export async function getAccountCoinAmount(args: {
-  aptosConfig: AptosConfig;
-  accountAddress: AccountAddressInput;
-  coinType?: MoveStructId;
-  faMetadataAddress?: AccountAddressInput;
-}): Promise<number> {
-  const { aptosConfig, accountAddress, coinType, faMetadataAddress } = args;
-
-  let coinAssetType: string | undefined = coinType;
-  let faAddress: string;
-
-  if (coinType !== undefined && faMetadataAddress !== undefined) {
-    faAddress = AccountAddress.from(faMetadataAddress).toStringLong();
-  } else if (coinType !== undefined && faMetadataAddress === undefined) {
-    // TODO Move to a separate function as defined in the AIP for coin migration
-    if (coinType === APTOS_COIN) {
-      faAddress = AccountAddress.A.toStringLong();
-    } else {
-      faAddress = createObjectAddress(AccountAddress.A, coinType).toStringLong();
-    }
-  } else if (coinType === undefined && faMetadataAddress !== undefined) {
-    const addr = AccountAddress.from(faMetadataAddress);
-    faAddress = addr.toStringLong();
-    if (addr === AccountAddress.A) {
-      coinAssetType = APTOS_COIN;
-    }
-    // The paired CoinType should be populated outside of this function in another
-    // async call. We cannot do this internally due to dependency cycles issue.
-  } else {
-    throw new Error("Either coinType, fungibleAssetAddress, or both must be provided");
-  }
-  const address = AccountAddress.from(accountAddress).toStringLong();
-
-  // Search by fungible asset address, unless it has a coin it migrated from
-  let where: any = { asset_type: { _eq: faAddress } };
-  if (coinAssetType !== undefined) {
-    where = { asset_type: { _in: [coinAssetType, faAddress] } };
-  }
-
-  const data = await getAccountCoinsData({
-    aptosConfig,
-    accountAddress: address,
-    options: {
-      where,
-    },
-  });
-
-  // commonjs (aka cjs) doesnt handle Nullish Coalescing for some reason
-  // might be because of how ts infer the graphql generated scheme type
-  return data[0] ? data[0].amount : 0;
 }
 
 export async function getAccountCoinsData(args: {